libs2eplugins: Requiem: intercept the mapped segments of the target process
The LinuxMonitor plugin has two specific signals
called "onProcessLoad" and "onMemoryMap".

We can subscribe to these two signals. When our target program
is loaded and running as a process, we record its pid.

Later on, whenever the kernel maps a segment for the target process,
we can record them and build its vmmap (which will be done in the next
commit).

Signed-off-by: Marco Wang <m.aesophor@gmail.com>